restructured packages for readability and clarity
