+++ .gitignore
@@ -2,7 +2,7 @@ junk.txt
 archive
 .env
 .docker_env
-
+junk
 # If you prefer the allow list template instead of the deny list, see community template:
 # https://github.com/github/gitignore/blob/main/community/Golang/Go.AllowList.gitignore
 #
