+++ api/handlers_api.go
@@ -5,14 +5,15 @@ import (
 	"log"
 
 	cerr "github.com/saeidalz13/battleship-backend/internal/error"
-	md "github.com/saeidalz13/battleship-backend/models"
+	mb "github.com/saeidalz13/battleship-backend/models/battleship"
+	mc "github.com/saeidalz13/battleship-backend/models/connection"
 )
 
 type RequestHandler interface {
-	HandleCreateGame() *md.Message[md.RespCreateGame]
-	HandleReadyPlayer() (*md.Message[md.NoPayload], *md.Game)
-	HandleJoinPlayer() (*md.Message[md.RespJoinGame], *md.Game)
-	HandleAttack() (*md.Message[md.RespAttack], *md.Player)
+	HandleCreateGame() *mc.Message[mc.RespCreateGame]
+	HandleReadyPlayer() (*mc.Message[mc.NoPayload], *mb.Game)
+	HandleJoinPlayer() (*mc.Message[mc.RespJoinGame], *mb.Game)
+	HandleAttack() (*mc.Message[mc.RespAttack], *mb.Player)
 }
 
 // Every incoming valid request will have this structure
@@ -40,23 +41,23 @@ func NewRequest(session *Session, payload ...[]byte) *Request {
 	return &wsReq
 }
 
-func (r *Request) HandleCreateGame() *md.Message[md.RespCreateGame] {
+func (r *Request) HandleCreateGame() *mc.Message[mc.RespCreateGame] {
 	game := r.Session.GameManager.AddGame()
 	r.Session.GameUuid = game.Uuid
 
 	hostPlayer := game.CreateHostPlayer(r.Session.ID)
 	r.Session.Player = hostPlayer
 
-	resp := md.NewMessage[md.RespCreateGame](md.CodeCreateGame)
-	resp.AddPayload(md.RespCreateGame{GameUuid: game.Uuid, HostUuid: hostPlayer.Uuid})
+	resp := mc.NewMessage[mc.RespCreateGame](mc.CodeCreateGame)
+	resp.AddPayload(mc.RespCreateGame{GameUuid: game.Uuid, HostUuid: hostPlayer.Uuid})
 	return &resp
 }
 
 // User will choose the configurations of ships on defence grid.
 // Then the grid is sent to backend and adjustment happens accordingly.
-func (r *Request) HandleReadyPlayer() (*md.Message[md.NoPayload], *md.Game) {
-	var readyPlayerReq md.Message[md.ReqReadyPlayer]
-	resp := md.NewMessage[md.NoPayload](md.CodeReady)
+func (r *Request) HandleReadyPlayer() (*mc.Message[mc.NoPayload], *mb.Game) {
+	var readyPlayerReq mc.Message[mc.ReqReadyPlayer]
+	resp := mc.NewMessage[mc.NoPayload](mc.CodeReady)
 
 	if err := json.Unmarshal(r.Payload, &readyPlayerReq); err != nil {
 		resp.AddError(err.Error(), cerr.ConstErrInvalidPayload)
@@ -71,13 +72,13 @@ func (r *Request) HandleReadyPlayer() (*md.Message[md.NoPayload], *md.Game) {
 
 	// Check to see if rows and cols are equal to game's grid size
 	rows := len(readyPlayerReq.Payload.DefenceGrid)
-	if rows != md.GameGridSize {
-		resp.AddError(cerr.ErrDefenceGridRowsOutOfBounds(rows, md.GameGridSize).Error(), cerr.ConstErrReady)
+	if rows != mb.GameGridSize {
+		resp.AddError(cerr.ErrDefenceGridRowsOutOfBounds(rows, mb.GameGridSize).Error(), cerr.ConstErrReady)
 		return &resp, nil
 	}
 	cols := len(readyPlayerReq.Payload.DefenceGrid[0])
-	if cols != md.GameGridSize {
-		resp.AddError(cerr.ErrDefenceGridColsOutOfBounds(cols, md.GameGridSize).Error(), cerr.ConstErrReady)
+	if cols != mb.GameGridSize {
+		resp.AddError(cerr.ErrDefenceGridColsOutOfBounds(cols, mb.GameGridSize).Error(), cerr.ConstErrReady)
 		return &resp, nil
 	}
 
@@ -88,9 +89,9 @@ func (r *Request) HandleReadyPlayer() (*md.Message[md.NoPayload], *md.Game) {
 
 // Join user sends the game uuid and if this game exists,
 // a new join player is created and added to the database
-func (r *Request) HandleJoinPlayer() (*md.Message[md.RespJoinGame], *md.Game) {
-	var joinGameReq md.Message[md.ReqJoinGame]
-	resp := md.NewMessage[md.RespJoinGame](md.CodeJoinGame)
+func (r *Request) HandleJoinPlayer() (*mc.Message[mc.RespJoinGame], *mb.Game) {
+	var joinGameReq mc.Message[mc.ReqJoinGame]
+	resp := mc.NewMessage[mc.RespJoinGame](mc.CodeJoinGame)
 
 	if err := json.Unmarshal(r.Payload, &joinGameReq); err != nil {
 		resp.AddError(err.Error(), cerr.ConstErrInvalidPayload)
@@ -108,14 +109,14 @@ func (r *Request) HandleJoinPlayer() (*md.Message[md.RespJoinGame], *md.Game) {
 	r.Session.GameUuid = game.Uuid
 	r.Session.Player = joinPlayer
 
-	resp.AddPayload(md.RespJoinGame{GameUuid: game.Uuid, PlayerUuid: joinPlayer.Uuid})
+	resp.AddPayload(mc.RespJoinGame{GameUuid: game.Uuid, PlayerUuid: joinPlayer.Uuid})
 	return &resp, game
 }
 
 // Handle the attack logic for the incoming request
-func (r *Request) HandleAttack() (*md.Message[md.RespAttack], *md.Player) {
-	var reqAttack md.Message[md.ReqAttack]
-	resp := md.NewMessage[md.RespAttack](md.CodeAttack)
+func (r *Request) HandleAttack() (*mc.Message[mc.RespAttack], *mb.Player) {
+	var reqAttack mc.Message[mc.ReqAttack]
+	resp := mc.NewMessage[mc.RespAttack](mc.CodeAttack)
 
 	if err := json.Unmarshal(r.Payload, &reqAttack); err != nil {
 		resp.AddError(err.Error(), cerr.ConstErrInvalidPayload)
@@ -124,7 +125,7 @@ func (r *Request) HandleAttack() (*md.Message[md.RespAttack], *md.Player) {
 
 	x := reqAttack.Payload.X
 	y := reqAttack.Payload.Y
-	if x > md.GameValidUpperBound || y > md.GameValidUpperBound || x < md.GameValidLowerBound || y < md.GameValidLowerBound {
+	if x > mb.GameValidUpperBound || y > mb.GameValidUpperBound || x < mb.GameValidLowerBound || y < mb.GameValidLowerBound {
 		resp.AddError(cerr.ErrXorYOutOfGridBound(x, y).Error(), cerr.ConstErrAttack)
 		return &resp, nil
 	}
@@ -142,7 +143,7 @@ func (r *Request) HandleAttack() (*md.Message[md.RespAttack], *md.Player) {
 	}
 
 	// Check if the attack position was already hit before (invalid position to attack)
-	if attacker.AttackGrid[x][y] != md.PositionStateAttackGridEmpty {
+	if attacker.AttackGrid[x][y] != mb.PositionStateAttackGridEmpty {
 		resp.AddError(cerr.ErrAttackPositionAlreadyFilled(x, y).Error(), cerr.ConstErrAttack)
 		return &resp, nil
 	}
@@ -166,14 +167,14 @@ func (r *Request) HandleAttack() (*md.Message[md.RespAttack], *md.Player) {
 	defender.IsTurn = true
 
 	// If the attacker missed
-	if positionCode == md.PositionStateAttackGridMiss {
+	if positionCode == mb.PositionStateAttackGridMiss {
 		// adjust attack grid for attacker
-		attacker.AttackGrid[x][y] = md.PositionStateAttackGridMiss
+		attacker.AttackGrid[x][y] = mb.PositionStateAttackGridMiss
 
-		resp.AddPayload(md.RespAttack{
+		resp.AddPayload(mc.RespAttack{
 			X:               x,
 			Y:               y,
-			PositionState:   md.PositionStateAttackGridMiss,
+			PositionState:   mb.PositionStateAttackGridMiss,
 			SunkenShipsHost: game.HostPlayer.SunkenShips,
 			SunkenShipsJoin: game.JoinPlayer.SunkenShips,
 		})
@@ -182,21 +183,21 @@ func (r *Request) HandleAttack() (*md.Message[md.RespAttack], *md.Player) {
 
 	// Apply the attack to the position for both defender and attacker
 	defender.HitShip(positionCode, x, y)
-	attacker.AttackGrid[x][y] = md.PositionStateAttackGridHit
+	attacker.AttackGrid[x][y] = mb.PositionStateAttackGridHit
 
 	// Initialize the payload
-	resp.AddPayload(md.RespAttack{
+	resp.AddPayload(mc.RespAttack{
 		X:             x,
 		Y:             y,
-		PositionState: md.PositionStateAttackGridHit,
+		PositionState: mb.PositionStateAttackGridHit,
 	})
 
 	// Check if the attack cause the ship to sink
 	if defender.IsShipSunken(positionCode) {
 		// Check if this sunken ship was the last one and the attacker is lost
 		if defender.IsLoser() {
-			defender.MatchStatus = md.PlayerMatchStatusLost
-			attacker.MatchStatus = md.PlayerMatchStatusWon
+			defender.MatchStatus = mb.PlayerMatchStatusLost
+			attacker.MatchStatus = mb.PlayerMatchStatusWon
 			game.FinishGame()
 		}
 	}
